Documentation

Documentation for JuliaGo.jl

Documentation

Types & Functions

Index

Types & Functions

MCTS

JuliaGo.Entry — Type

Entry

Type storing the wins, draws and number of visits for a Node.

JuliaGo.Node — Type

Node

Type used to represent nodes on the game tree.

JuliaGo.MCTS! — Method

MCTS!(b::Board, P::Node, num_rollouts::UInt64)

Monte Carle Tree Search algorithm. Returns the Node with the highest visit count via robust_child.

JuliaGo.backpropagate! — Method

backpropagate!(C::Node, b::Board, results::Dict{PieceColor,UInt64})

Backpropagates the result of a rollout up the tree until the root node. The root node is defined by root.parent == nothing. Returns the root node.

JuliaGo.expansion! — Method

expansion!(b::Board, P::Node, num_rollouts::UInt64)::Node

Selects a child node at random and removes the move from the parent node's unexplored_moves list. Conducts rollouts from the child node until a terminal state and backpropagates the result until the root node of of the tree (P is not the root node). This rollout function is multi-threaded i.e. each rollout is perfomed by a different thread. The results are collected and counted (making it all thread safe).

JuliaGo.firstpass! — Method

 firstpass!(b::Board, P::Node, num_rollouts::UInt64)

Given the root node of tree, where no simulation have yet occured, populate all the children of the root node with a value based on rollouts and back propagation.

JuliaGo.robust_child — Method

robust_child(P::Node)::Move

Selects the child with the highest visit count N and returns the Move associated with that child. This is the default method of choosing the action but others exist, see secure child and max child.

JuliaGo.rollout — Method

rollout(b::Board, policy)::PieceColor

Plays moves according to a policy, a neural network, until a terminal state is reached. When a terminal state is reached the colour of the winning side is returned, COLOR_NONE if it is a draw

JuliaGo.rollout — Method

rollout(b::Board)::PieceColor

Plays moves randomly until a terminal state is reached.

JuliaGo.selection_policy — Method

function selection_policy(C::Node)

Calculates the Upper Confidence Bound for Trees (UCT) via

$\mathrm{UCT} = \frac{w_i}{n_i} + C\sqrt{\frac{\ln N_i}{n_i}}$

where

$w_i$ is the number of wins associated with taking the action that results in this node i.e. the number of wins from the parent's perspecive, $n_i$ is the number of simulations from this node, $C = \sqrt{2}$ is a coefficient controlling the balance between exploration and exploitation, $N_i$ is the number of simulations of the parent node.

Neural Network

JuliaGo.nn_input — Method

nn_input

Convert a board to a (8, 8, 17) Array{Float32}, suitable for input to the neural network. There are 12 (8, 8) arrays for the pieces, 4 for the castling rights and 1 for the colour of the side to move.

JuliaGo.nn_output — Method

nn_output(b::Board, next_board::Board)::Array{Float32,3}

Given a board b and the next_board, convert the move to a (8, 8, 73) array. The 73 values are encoded as follows: 1-7 horizontal right, 8-14 horizontal left, 15-21 vertical up, 22-28 vertical down, 29-35 north east, 36-42 north west, 42-49 south east, 50-56 south west, 57-64 = knight moves, 65-67 pawn mv NE and underpromote to bishop, knight, rook, 68-70 pawn mv NW, 71-73 pawn mv N.

JuliaGo.output_to_move — Method

output_to_move

Convert the index of the highest probability element in the model output array to a move.

JuliaGo.parse_pgns — Method

parse_pgns

Parse a list of pgns and return arrays for training the neural network. Positions in each game are selected according with probability prob to prevent highly correlated positions.

JuliaGo.predict_move — Method

predict_move

Given a board and policy, output the array of possible next moves with associated probabilty. If the policy outputs and illegal move, a legal move is selected at random.

JuliaGo.train_test_data — Method

input(preloaded, ratio)

Create train and test sets from the pgns, split according to the ratio.

JuliaGo.ResNetBlock — Method

ResNetBlock()

Defines a single ResNet block - two rectified batch-normalized convolutional layers with a skip connection.

JuliaGo.batch_train! — Function

batch_train!(loss,
ps,
train_x,
train_y,
test_x,
test_y,
opt,
model,
batchsize = 64,
epochs = 100000,)

At each epoch, a random selection of the train_x data of size batchsize is trasferred to the gpu and the model trained - this is mini-batch gradient descent. The model is saved in cpu format at every 5% of the total number of epochs.

JuliaGo.policy_network_train — Function

policy_network_train(preloaded_data=false, model_name="policy_network.bson")

Train a policy network on a dataset of professional chess games. The aim of the policy network is to predict the move from a state that a professional would likely play.

By default the model is transferred to the gpu.

Pass preloaded_data=true if the pgns have already been parsed and converted into a input for the neural network (considerable time saving).

Index

JuliaGo.Entry

JuliaGo.Node

JuliaGo.MCTS!

JuliaGo.ResNetBlock

JuliaGo.backpropagate!

JuliaGo.batch_train!

JuliaGo.expansion!

JuliaGo.firstpass!

JuliaGo.nn_input

JuliaGo.nn_output

JuliaGo.output_to_move

JuliaGo.parse_pgns

JuliaGo.policy_network_train

JuliaGo.predict_move

JuliaGo.robust_child

JuliaGo.rollout

JuliaGo.rollout

JuliaGo.selection_policy

JuliaGo.train_test_data